moves haversine function to its own file which is refered in __init__
